feat: exhaustive set of tests for all verbs
