+++ packages/at_secondary_server/lib/src/connection/inbound/connection_util.dart
@@ -189,70 +189,6 @@ class InboundIdleChecker {
   }
 }
 
-enum AtVerb {
-  // Authenticated
-  update(requiresAuth: true),
-  llookup(requiresAuth: true),
-  delete(requiresAuth: true),
-  sync(requiresAuth: true),
-  notify(requiresAuth: true, hasSubcommands: true),
-  monitor(requiresAuth: true),
-  otp(requiresAuth: true),
-  keys(requiresAuth: true, hasSubcommands: true),
-  batch(requiresAuth: true),
-  stats(requiresAuth: true),
-  config(requiresAuth: true),
-
-  // Unauthenticated
-  from(requiresAuth: false),
-  cram(requiresAuth: false),
-  pkam(requiresAuth: false),
-  pol(requiresAuth: false),
-  scan(requiresAuth: false),
-  lookup(requiresAuth: false),
-  plookup(requiresAuth: false),
-  info(requiresAuth: false),
-  noop(requiresAuth: false),
-
-  // available both ways, must parse subcommands
-  enroll(requiresAuth: false, hasSubcommands: true);
-
-  const AtVerb({required this.requiresAuth, this.hasSubcommands = false});
-  final bool requiresAuth;
-  final bool hasSubcommands;
-
-  static AtVerb? tryParse(String value) =>
-      AtVerb.values.where((v) => v.name == value).firstOrNull;
-}
-
-// all subcommands typically require authentication
-// there is some overlap in subcommands
-enum Subcommand {
-  //enroll
-  request(requiresAuth: false),
-  approve,
-  deny,
-  revoke,
-  unrevoke,
-  delete,
-  list,
-  fetch,
-  //notify
-  remove,
-  status,
-  update,
-  all,
-  //keys
-  put,
-  get;
-
-  const Subcommand({this.requiresAuth = true});
-  final bool requiresAuth;
-
-  static Subcommand? tryParse(String value) =>
-      Subcommand.values.where((v) => v.name == value).firstOrNull;
-}
-
 class InboundCommandValidator {
   /// This function validates a command on a connection. The criteria is the following:
   /// 1. checks if connection is invalid, closing the connection if requires
@@ -270,7 +206,7 @@ class InboundCommandValidator {
           'Connection is invalid, closing connection');
     }
 
-    // why does scan delimate with a space....
+    // why does scan delimit with a space....
     if (command.contains('scan ')) {
       return;
     }
@@ -292,7 +228,7 @@ class InboundCommandValidator {
         (throw InvalidSyntaxException(
             'Received invalid verb that does not match protocol spec'));
 
-    // determine auth requirement - may be overriden if verb has subcommands
+    // determine auth requirement - may be overridden if verb has subcommands
     bool requiresAuth = verb.requiresAuth;
     if (verb.hasSubcommands) {
       String rawSubcommand = command.split(":")[1].trim();
