+++ packages/at_secondary_server/test/inbound_message_listener_test.dart
@@ -1,0 +1,260 @@
+import 'dart:async';
+import 'dart:convert';
+
+import 'package:at_commons/at_commons.dart';
+import 'package:at_secondary/src/connection/inbound/connection_util.dart';
+import 'package:at_secondary/src/connection/inbound/inbound_connection_metadata.dart';
+import 'package:at_secondary/src/connection/inbound/inbound_message_listener.dart';
+import 'package:at_secondary/src/server/at_secondary_impl.dart';
+import 'package:at_server_spec/at_server_spec.dart';
+import 'package:mocktail/mocktail.dart';
+import 'package:test/test.dart';
+
+import 'test_utils.dart';
+
+StreamController<String> sc = StreamController<String>();
+
+void streamCallBack(List<int> data, InboundConnection connection) {}
+Future<void> callback(String value, InboundConnection connection) async {
+  sc.add(value);
+}
+
+void _expectCommandsToValidate(
+    List<String> commands,
+    InboundConnection connection,
+    ) {
+  for (final command in commands) {
+    expect(
+          () => InboundCommandValidator.validate(
+        utf8.encode(command).toList(),
+        connection,
+      ),
+      returnsNormally,
+      reason: 'Expected valid command: $command',
+    );
+  }
+}
+
+void _expectCommandsToRequireAuth(
+    List<String> commands,
+    InboundConnection connection,
+    ) {
+  for (final command in commands) {
+    expect(
+          () => InboundCommandValidator.validate(
+        utf8.encode(command).toList(),
+        connection,
+      ),
+      throwsA(isA<UnAuthenticatedException>()),
+      reason: 'Expected auth-required command: $command',
+    );
+  }
+}
+
+void main() async {
+  late FakeSocket socket;
+  late InboundConnection connection;
+  late FakeInboundConnection fakeConnection;
+
+  late InboundConnectionMetadata authenticatedMetadata;
+  late InboundConnectionMetadata unAuthenticatedMetadata;
+
+  setUpAll(() {
+    verbTestsSetUpLogging();
+    authenticatedMetadata = InboundConnectionMetadata()..isAuthenticated = true;
+    unAuthenticatedMetadata = InboundConnectionMetadata();
+  });
+
+  setUp(() async {
+    sc = StreamController<String>();
+    socket = FakeSocket();
+    connection = MockInboundConnection();
+    fakeConnection = FakeInboundConnection(socket, authenticatedMetadata);
+    when(() => connection.close()).thenAnswer((_) async {});
+    when(() => connection.isInValid()).thenReturn(false);
+
+    final mockManager = MockInboundConnectionManager();
+    final mockPool = MockInboundConnectionPool();
+    when(() => mockManager.pool).thenReturn(mockPool);
+    when(() => mockPool.remove(fakeConnection)).thenReturn(null);
+    AtSecondaryServerImpl.getInstance().inboundConnectionManager = mockManager;
+  });
+
+  // testing InboundConnectionValidator.validate(), mainly focusing on what it throws and when
+  // happy path of this group is covered by inbound message listener.
+  group('A test to verify commands are properly invalidated in the buffer', () {
+    test('validate a successful command -> should pass', () {
+      when(() => connection.metaData).thenReturn(authenticatedMetadata);
+      expect(
+            () => InboundCommandValidator.validate(
+            utf8.encode('lookup:public:publickey@alice\n').toList(),
+            connection),
+        returnsNormally,
+      );
+    });
+
+    test('validate a failed verb (length > 32)', () {
+      when(() => connection.metaData).thenReturn(authenticatedMetadata);
+      expect(
+            () => InboundCommandValidator.validate(
+            utf8.encode('lookupdjflsdfspublickeylice:').toList(), connection),
+        throwsA(isA<InvalidSyntaxException>()),
+      );
+    });
+
+    test('validate a failed verb (trying to run authenticated verb)', () {
+      when(() => connection.metaData).thenReturn(unAuthenticatedMetadata);
+      expect(
+            () => InboundCommandValidator.validate(
+            utf8.encode('monitor\n').toList(), connection),
+        throwsA(isA<UnAuthenticatedException>()),
+      );
+    });
+
+    test('validate all unauthenticated top-level verbs -> should pass', () {
+      when(() => connection.metaData).thenReturn(unAuthenticatedMetadata);
+      _expectCommandsToValidate(unauthenticatedTopLevelCommands, connection);
+    });
+
+    test('validate all authenticated top-level verbs unauthenticated -> fail',
+            () {
+          when(() => connection.metaData).thenReturn(unAuthenticatedMetadata);
+          _expectCommandsToRequireAuth(authenticatedTopLevelCommands, connection);
+        });
+
+    test('validate all authenticated top-level verbs authenticated -> pass',
+            () {
+          when(() => connection.metaData).thenReturn(authenticatedMetadata);
+          _expectCommandsToValidate(authenticatedTopLevelCommands, connection);
+        });
+
+    test('validate subcommand auth requirements for notify/keys/enroll', () {
+      when(() => connection.metaData).thenReturn(unAuthenticatedMetadata);
+      _expectCommandsToValidate(subcommandsAllowedWithoutAuth, connection);
+      _expectCommandsToRequireAuth(subcommandsRequiringAuth, connection);
+    });
+
+    test('validate scan with trailing arguments bypasses verb parsing', () {
+      when(() => connection.metaData).thenReturn(unAuthenticatedMetadata);
+      expect(
+            () => InboundCommandValidator.validate(
+          utf8.encode('scan this-is-accepted\n').toList(),
+          connection,
+        ),
+        returnsNormally,
+      );
+    });
+
+    test('invalidate bad connection', () {
+      when(() => connection.metaData).thenReturn(unAuthenticatedMetadata);
+      when(() => connection.isInValid()).thenReturn(true);
+      expect(
+            () => InboundCommandValidator.validate(
+            utf8.encode('monitor\n').toList(), connection),
+        throwsA(isA<ConnectionInvalidException>()),
+      );
+    });
+
+    test('invalidate (non authenticated junk)', () {
+      when(() => connection.metaData).thenReturn(unAuthenticatedMetadata);
+      when(() => connection.isInValid()).thenReturn(true);
+      expect(
+            () => InboundCommandValidator.validate(
+            utf8
+                .encode(
+                'monitoradshfajdsfkjalsdkjflaksdjlfkajsdlkfjalksdjflaksdjflaksdjflkadsjflkasdjflkajsdlkfjasldkfjalksdjflaksdjflaksdjfalskdfjalsdkfjalksdjfalksdjflaksdjflaksdjflaksdjflakdsjflakdsjfalksdjflaksdjflaksdjflaksdjflaksdjflaksdjflaksdjflaksdjflaksdjflaksdjflaksdjfalskdjfalskdjfalksdjfalskdjfalksdjflaskdjfalksdjfalskdjfalsdkfjalsdkjf')
+                .toList(),
+            connection),
+        throwsA(isA<ConnectionInvalidException>()),
+      );
+    });
+  });
+
+  group(
+      'A test to verify that the message listener is properly handling commands in the buffer',
+          () {
+        test('validate one-word commands -> should pass', () async {
+          var listener = InboundMessageListener(fakeConnection);
+          final streamFuture = expectLater(
+              sc.stream,
+              emitsInOrder([
+                equalsIgnoringCase('scan'),
+                equalsIgnoringCase('info'),
+              ]));
+          listener.listen(callback, streamCallBack);
+          fakeConnection.socket.addData('scan\n');
+          fakeConnection.socket.addData('info\n');
+          await streamFuture;
+        });
+
+        test('validate a colon separated command -> should pass', () async {
+          var listener = InboundMessageListener(fakeConnection);
+          var lookup = 'lookup:public:publickey@alice:metadata:ttl:0\n';
+          var from = 'from:@mchicken\n';
+          var pkam = 'pkam:superadvancedcoolsignature\n';
+          final streamFuture = expectLater(
+              sc.stream,
+              emitsInOrder([
+                equalsIgnoringCase(lookup.trim()),
+                equalsIgnoringCase(from.trim()),
+                equalsIgnoringCase(pkam.trim()),
+              ]));
+          listener.listen(callback, streamCallBack);
+          fakeConnection.socket.addData(lookup);
+          fakeConnection.socket.addData(from);
+          fakeConnection.socket.addData(pkam);
+          await streamFuture;
+        });
+
+        test('validate a colon & json included command -> should pass', () async {
+          var listener = InboundMessageListener(fakeConnection);
+          listener.listen(callback, streamCallBack);
+          var enroll =
+              'enroll:request:{"appName":"wavi","deviceName":"iphone","namespaces":{"wavi":"rw"},"otp":"<otp>","apkamPublicKey":"<apkamPublicKey>","encryptedAPKAMSymmetricKey": "<encryptedAPKAMSymmetricKey>"}\n';
+          final streamFuture = expectLater(
+              sc.stream,
+              emitsInOrder([
+                equalsIgnoringCase(enroll.trim()),
+              ]));
+          fakeConnection.socket.addData(enroll);
+          await streamFuture;
+        });
+
+        test('partial command validation -> should pass', () async {
+          // receive command in increments, expect to validate by packet2
+          // shouldn't fail and should emit one command.
+          when(() => connection.metaData).thenReturn(authenticatedMetadata);
+          var listener = InboundMessageListener(fakeConnection);
+          listener.listen(callback, streamCallBack);
+          var enroll =
+              'enroll:request:{"appName":"wavi","deviceName":"iphone","namespaces":{"wavi":"rw"},"otp":"<otp>","apkamPublicKey":"<apkamPublicKey>","encryptedAPKAMSymmetricKey": "<encryptedAPKAMSymmetricKey>"}\n';
+          final packet1 = enroll.substring(0, 5);
+          final packet2 = enroll.substring(5, 40);
+          final packet3 = enroll.substring(40);
+          final streamFuture = expectLater(
+              sc.stream,
+              emitsInOrder([
+                equalsIgnoringCase(enroll.trim()),
+              ]));
+          fakeConnection.socket.addData(packet1);
+          fakeConnection.socket.addData(packet2);
+          fakeConnection.socket.addData(packet3);
+          await streamFuture;
+        });
+
+        test('validate representative command for each verb -> should pass',
+                () async {
+              var listener = InboundMessageListener(fakeConnection);
+              listener.listen(callback, streamCallBack);
+              final streamFuture = expectLater(
+                  sc.stream,
+                  emitsInOrder(representativeListenerCommands
+                      .map((command) => equalsIgnoringCase(command.trim()))
+                      .toList()));
+              for (final command in representativeListenerCommands) {
+                fakeConnection.socket.addData(command);
+              }
+              await streamFuture;
+            });
+      });
+}
